Refactor workdir methods to new package
